requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-on-heroku==1.0.1
djangorestframework==3.11.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.5
Pygments==2.6.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
